requirements.txt
channels-redis==3.2.0
asgiref==3.3.1
autobahn==20.7.1
backports.ssl-match-hostname==3.7.0.1
channels==3.0.2
daphne==3.0.1
Django==3.1.4
django-import-export==2.4.0
h5py==3.1.0
msgpack-python==0.5.6
PyMySQL==0.10.1
nltk==3.5
numpy==1.19.2
pika==1.1.0
Pillow==8.0.1
psycopg2==2.8.5
pytorch==1.7.0
torchaudio==0.7.0
torchvision==0.8.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
scipy==1.5.2
six==1.15.0
tqdm==4.51.0
Twisted==20.3.0
txaio==20.4.1
uWSGI==2.0.19.1
websocket-client==0.57.0
zope.interface==5.2.0
boto==2.49.0
scikit-learn==0.23.2
